changelog: fix undefined variable
Summary:
This is causing the Mononoke and hg tests to break on an undefined variable.
This looks like it might have been a refactoring accident when the code
defining this got extracted into D22240176 (but the code using it landed in
D22240177 (https://github.com/facebookexperimental/eden/commit/ec58e7290322a46cef20406aa6e825ead4718b92)).

This diff unbreaks the tests by defaulting the parameter to False (which
seems coherent with the idea of that latter diff), and puts it in a place
that'll trigger a merge conflict for quark-zju when he rebases D22240176.

Reviewed By: farnz

Differential Revision: D22408588

fbshipit-source-id: 496808742a13dfeb17989123742a0aa8bae17b38