http_client: add buffering to CborStream
Summary:
As was pointed out in the review for D22280745 (https://github.com/facebookexperimental/eden/commit/d73c63d8620f8e1e7fccd663ba02412514b84861), `CborStream` is inefficient in situations where the underlying stream produces chunks that are much smaller than the size of the serialized items. To avoid pathological behavior, make `CborStream` buffer the incoming data, and only attempt deserialization if enough data has accumulated.

For now, the buffer size is fixed (with a default of 1MB, chosen arbitrarily). In the future, it might make sense to have the stream adjust the buffer size based on the average size of observed deserialized values.

Reviewed By: quark-zju

Differential Revision: D22370164

fbshipit-source-id: ed940c56ca2cbbfc07f01d47becf6f1d71872872