revlog: bypass some transaction logic when adding commits with modern setup
Summary:
With narrow-heads, visible heads are explicitly controlled by commit
references. Adding commits can be just writing them out directly.

This mainly removes the "buffered" writes of `00changelog.i`.
Instead of writing pending changes to `00changelog.i.a`, they
are directly written to `00changelog.i` (or buffered in memory
with future changes).

This does not bypass all transaction logic. Truncation can still
happen. Strip is also unaffected.

The change is incomplete. In the future, pending changes will
be written in-memory to the Rust HgCommits struct and we no
longer write directly to revlog.

Reviewed By: DurhamG

Differential Revision: D22240176

fbshipit-source-id: ac9d20ab95ff304fb285a503d2d3db815942d5b3