edenapi: reuse HttpClient
Summary:
D22396026 made it so that `HttpClient::send_async` no longer consumes `self`. This means that instead of creating a new HTTP client for each request, we can reuse the same one.

This has the benefit of allowing for connection reuse (which was the point of D22396026), resulting in lower latency for serial fetches.

Reviewed By: quark-zju

Differential Revision: D22397768

fbshipit-source-id: 9d066c1ec64a6aa1b36ec674ef294030c1f90b41