pyrevisionstore: add EdenAPI store bindings
Summary: Add add a `edenapistore` class to that wraps a `EdenApiHgIdRemoteStore`. This class is purely used as a means to set up the stores from Python code, and is only used as a way to get an `Arc<EdenApiHgIdRemoteStore>` to the Rust content store. It has no functionality of its own.

Reviewed By: quark-zju

Differential Revision: D22449702

fbshipit-source-id: ad2094c79da523071b6ed8344c8dde706e448c95