scs_server: compute descendant bookmarks in parallel
Summary:
Computing descendant bookmarks is done async-but-serially (i.e. in a single
tokio Task).  That means it is slow for repos with many bookmarks if you make a
request for an early commit.

This change makes it use multiple tokio Tasks, each one computing a batch of 100
bookmarks, giving true parallelism.

Reviewed By: krallin

Differential Revision: D22461931

fbshipit-source-id: a8908f7c20173c61b83d69c9dc37a5275937e2dc