mononoke: use bulkops in skiplist builder
Summary:
We already have a function to fetch all public changesets - let's use it
instead of re-implementing it.

The small caveat is that function in skiplist subcommand fetched all the
changesets (i.e. public and draft), so using bulkops function looks like a change in
behaviour. However it's actually the same - we index only public changesets for
skiplists anyway.

Reviewed By: krallin

Differential Revision: D22499940

fbshipit-source-id: ac8ad7d2b6ff0208e830a344877d7d2e93693abc