configs: avoid kicking off background generation twice
Summary:
When Mercurial detects that the current configuration is old, it kicks
off a background command to regenerate the configs. Because Mercurial loads
configs twice (once in the early stages of dispatch before the repo is loaded,
and once as the repo is instantiated), we can end up kicking off the backgroud
generation twice, which could result in two requests to the remote server.
Doubling our QPS.

Let's prevent kicking off a background generation twice.

The use of a global variable here is awkward, but I couldn't think of a great
way to track this information. I could create a file and use the mtime to track
the last time we kicked off a generation, but that felt like more complexity
than this should have. In the long run we should make Mercurial only load the
configs once.

Reviewed By: quark-zju

Differential Revision: D22507796

fbshipit-source-id: 101e0d3cdccfdaf62f7be60bc9480e6e25b43c88