mononoke_hg_sync_job: vendor sendunbundlereplay
Summary:
Vendor sendunbundlereplay instead of getting it from hg. The sync job is the
only thing that uses this, and all the test coverage for this is in Mononoke,
so it being an extension in hg isn't particularly helpful (and yet comes at the
cost of being harder to hotfix).

NOTE: this also includes a fix by StanislavGlebik to base64-encode the bookmark name.

Reviewed By: farnz

Differential Revision: D22513868

fbshipit-source-id: 03c941564255e18ef78ccddef826d26afd0e5d32