scs_server: increase list_bookmarks limits (part 1)
Summary:
Increase the max limits for repo_list_bookmarks and commit_list_descendant_bookmarks
from 1,000 to 10,000.  The higher number is still reasonable for a single request,
and this reduces the number of round-trips when there are lots of bookmarks.

This updates the server.  A later diff will increase the constant so that clients
can make use of it, but this diff must be landed and deployed before that change
can be made.

Reviewed By: mitrandir77

Differential Revision: D22524892

fbshipit-source-id: c216f4ba7fa60774990d87747c9d8ea9d551dc85