Deprecate setMinCompressBytes on Thrift Server (4/n)
Summary:
Deprecate server-side minCompressBytes.

Background:
The minCompressBytes on ThriftServer is another knob to control whether or not to compress payloads. The reason to deprecate it is twofold:
(1) to provide a single place for compression control.
(2) 99.99% of services only copy-paste the default value (200) anyways.

Since rocket transport will be the default transport for all future thrift RPC, D21702725 introduced the ability for client to pass CompressionConfig to server (in rocket) which informs server both codec and size limit. This way the compression is enabled and controlled only from the client-side.

Reviewed By: iahs

Differential Revision: D22354858

fbshipit-source-id: dd920d4d38969f8ee7894b1aaafbe71c0e8b8d86