tests: do not use `hg files` to list every file in a repo
Summary:
The `helpers-testrepo.sh` runs something like

  hg files --cwd /path/to/fbsource/fbcode/eden/scm/tests/

to check if `hg` works or not. Previously it fails fast because of the
lz4revlog requirement. The next diff will enable lz4revlog that makes
`hg files` work.

That `hg files` command (if does not fail) is quite expensive because it has no
patterns and lists every file.

It can also be undesirable if the in-repo hg does some kind of auto migration
that makes the repo unusable by the system hg. Therefore just unconditionally
use the system hg.

Reviewed By: DurhamG

Differential Revision: D22419967

fbshipit-source-id: 2d5c75696efecce1ec38a2e4fa23aff101219545