mononoke/lfs_server: popularity: add ODS counters, timeouts and tests
Summary:
ODS counters are helpful to know if the feature is turned on or off without
requiring a traffic spike, so let's log them. Also, let's add timeouts in here,
so we know if things aren't working as expected (I did check in the Mononoke
LFS dataset — 10ms is a very conservative number, that's way beyond the p99 of
batch requests, which include potentially many counter checks).

To make this easier to iterate on, let's also add tests.

Reviewed By: StanislavGlebik

Differential Revision: D22545853

fbshipit-source-id: 02ea4484a4e4ba0dfd4a71030c129eb5c6bb1ec9