scs_server: increase list_bookmarks limits (part 2)
Summary:
Increase the max limits for repo_list_bookmarks and commit_list_descendant_bookmarks
from 1,000 to 10,000.  The higher number is still reasonable for a single request,
and this reduces the number of round-trips when there are lots of bookmarks.

This updates the constants in the thrift definition.  The server was updated to accept
higher limits in a previous diff.

Reviewed By: mitrandir77

Differential Revision: D22524891

fbshipit-source-id: bc47c1e50af35f213ebcfbca1574669e79b2fe92